Move `CacheDecoder` to `rustc_query_impl`
